Interact with shadow detections
Up to now, dragging a shadow detection, or setting it to obscured would not work, because the updateAnnotation call to the backend would fail (the annotation does not actually exist in the db after all so it cannot be updated). This code change checks for shadow status when calling updateAnnotation and calls createAnnotation instead if we are dealing with a shadow. Any kind of interaction with a shadow (dragging, changing detection type or changing visibility status) will instantly convert the shadow to an actual detection. The only thing that you cannot do with shadows that you can do with actual detections is to delete them. If you want to get rid of the shadow you have to delete the annotation that is casting that shadow.