Keep consistent identicons whenever possible.
We rerun the optimal assignment algorithm after every change to the annotations, because in general we have to assume that assignments may change when the input changes. However the proposed id in cases where neither of the pair's components allready have an objectId used to generated with crypto.randomUUID. This meant that all identities that have not been committed would change every time we recompute. This can be visually jarring. To try and minimise unneccesary changes, we now assign the ids of one of the component parts as the objectId whenever possible.