Fixed issue TS2307 that interfered with deployment
We've been struggling with TS2307 for a while now. Specifically errors of the form:
Could not resolve "$amplify/env/add-user-to-group"
What made this especially tricky is that this error seems top pop up in a variety of contexts. In some cases the backend build fails, and in some cases it is the frontend. this affects some people in their sandboxes and others (like us seemingly) only when they try to deploy. There is a huge range of "fixes" out there that people calim worked for them. We tried a lot of these with no joy. Ultimately we committed the autogenerated code in .amplify out of desperation. today I was working on something unrelated when I managed to get the same problem in a sandbox deploy (by deleting the entire .amplify folder in an attempt to force a fresh sandbox). Once I got there it quickly became apparent that the bug was actually quite stupid
import { env } from "$amplify/env/add-user-to-group"
is a valid line of code, but only inside:
amplify\functions\add-user-to-group\handler.ts
as this file will get processed AFTER
amplify\functions\add-user-to-group\resource.ts

During copy paste we ended up with that line inside some other function handlers. Everything still worked because the exported envs were similar enough for our purposes and the file "$amplify/env/add-user-to-group" had allready been generated in a previous run. But if you start a fresh deploy then this will only work if the add-user-to-group function is built before all the functions that erroneously try to import its config. Frustratingly, it seems that the build order of amplify functions is non-deterministic so it would sometimes work at random times, leading us to believe that whatever we did previously had solved the problem, only for the problem to resurface on the next clean build. I am reasonably confident that it is actually fixed now (but can only be 100% sure after a succesful branch deploy) so I have removed the .amplify/generated folder from git and added it back to .gitignore.