Attempted fix of timeout issue on tiles
I still get tiles that fail to load if I left the site open for a long time. My best guess is that some URLs are cached even though we never retrieved them. I've adapted the storagelayer to also store the timestamp at which a URL will expire and when we retrieve a URL from cache we first check that it is still valid for at least 30 seconds before attempting to use it.