Rotation detetion bugfix
The existing code assumed the presence of an Orientation field in the exif data that could be used to detect an image that should be rotated by 90/270 degrees (i.e. we must swap width and height dimensions). This tag should always be present in images from any remotely recent camera. However, in some cases image postprocessing software may rotate tthe image to the correct orientation and then remove the tag. In this case our existing code would fail in the wrong direction and always detect rotation rather than never detecting rotation. This commit fixes that.