Reload project management components when project is switched
I saw that if I select an imageset and define a spatial subset, then toggle to another project, select another imageset and go define another spatial subset, then both imagesets are active inside the "Define Sptiotemporal subsets" modal.
This is because both sets are listed in selectedSets, so the Spatial subset modal is actually doing the correct thing. We could add an effect to ImageSubsetManagement that watches the current project and resets all states when it changes, but we would have to do the same thing in all the other menegement components and keep these effects updated. The better solution is to force an actual reload (rather than simulating one) each time the project cahnges (because we never want to transfer state between projects). Tis is easily achieved by giving each of the menegement components a key that includes the current project id.