Example script for making bulk changes to db
I needed a single use script to swap width and height fields for images in the Addo set due to the bug that was fixed in 930e3f4da22f73a5206602cfa7f63ac2ef7ede6d. This was much more painful than anticipated, so I decided to commit the script so that an example exists, if anything like this ever needs to be done again.