Memoization of Showmarkers
It looks like the lag issue is back. I've memoized ShowMarkers (this was a big part of the solution before but was done on the consumer (of the ShowMarkers component side) and regressed. Now putting the memoization inside showmarkers itself which should be less likely to get dropped by an overeager updater such as myself.