Fixed lag error
The lag error resurfaced. As before it was mainly due to excessive rerenders of the MapContainerComponent. This component should really only render once, but due to the way we provide ShowMarkers, Legend etc as children to a higher level  prop is is actually quite tricky to avoid rerenders and assure that they stay avoided. Any property or child that is added carelessly at a complete;y different level breaks memoization and because it is only a problem when working with large herds this regression is only discovered days or weeks later, which makes it much harder to fix. We need to look at changing the design or implementation so we are more robust to changes such as these or alternatively to implement a regression test that runs automatically as part of CI/CD pipeline to prevent regressions from being pushed.