Centralized Error reporting and concurrency limit
This is a fairly big refactor where I added a new file limitedClient. This file contains code to create an Amplify GraphQL client and then wrap it in a decorator that:
 - applies a pLimit object that limits concurrency to 15 promises at a time to every async method provided by the client.
 - Checks the return value for an errors field and throws a GraphQLError if it exists. This is required for the rollback part of optimistic updating as implemented using tanstack query to work. It was added in a somewhat hitt and miss way before, but now it is per-definition active on all graphql methods.
 - Apply the same decorator to the return values of async calls. This is to ensure that lazy loading methods for resolving relationships will be subject to the same error-checking and concurrency-limitations invariants as all the static API methods.

I also did a little cleanup to the GlobalContext which is used to supply the limitedClient object to all components which use it.