Revert to autozoom by default
The zoom experiment from commit 78f4fd6f33a1d41da050b706458c02c13cd5f914 was a success, but for now we need to revert to having autoscale as the default. If a fixed zoom level is required, that can be specified in the job message. Currently we don't have UI elements to specify that. It will be added in a future commit.