Track loading state of BaseImage
Rewrote StorageLayer to properly inherit behaviours from other react-leaflet LayerComponents. Specifically we need the evented behaviour because we need to be able to set eventhandlers on the layer, specifically an handler for the "load" event which is fired when all visible tiles have been loaded. We also needed to delay the sending of the done event on each tile until the tile has actually been loaded (otherwise the GridLayer would emit the load event early).

Lastly I added a callback to the StorageLayer where it is used in BaseImage and made the BaseImage hide itself until the done event is fired. This means that you should never toggle to a next image and see tiles still loading. The longer term goal is to use this information to control sequential preloading of images (so image 2 doesn't start loading before image 1 is fully loaded etc.)