Various improvements to updateStats
It seems that the remaining leaderboard issues can be traced to lambdas timing out and then being retried causing the same graphql update to be applied multiple times. I've changed the code structure to first accumulate all updates for the PK and then apply all the updates in parallel. This should buy us at least an order of magnitude scalewise.